Fix edit button for non-existant pages
The logic for displaying the edit button on pages was: is the user in a
group that can edit the page, and can the user edit this specific page.
Now when a page does not exist, the user might have rights to edit
pages, but no rights to edit this specific page exist, so the logic
fails.

This is fixed by changing the logic to: is the user in a group that can
edit the page, and is this a non-existant page or can the user edit this
specific existant page.

Shit wooorks.