Added a language switcher support
The language of the user is switched to the selected view, restricted by
config. If the current user is anonymous (not logged in) then a client
side cookie is set with the language in it.

On request the locale from the user is used, if the current user is
anoymous then the language from the set cookie is taken. If no cookie is
found, the best matching language from the config is taken based on
request headers.

Locale selection function is now also available in templates to be able
to show the current used language in the base template.