Updated database model form dual language implementation
Column for name and description are duplicated with a prefix for Dutch
and English language.

All strings are now babel lazy evaluated strings, causing the to be
looked up on request in the .po files.

Removed the separation of timedelta calculation and formatation of it.
Reducing the number of nested function calls for one line of code.

On loading of the Activity model the name and description properties are
set based on cookie, headers or saved locale. Those properties were used
everywhere in viaduct.

If set language is unavailable for the current activity, and suffix is
added to notify the user of content in a different language.