Merged single scripts into manage.py and added rules for other tasks.
Moved the create_flaskjs 3-liner script into manage.py as the jade task
and updated watch.sh accordingly. Also did the same for the creation of
the flask.js file.

Moved the test_modules into a test directory and created a rule in manage.py to
run all tests found in this directory.

Added a task into manage.py to run the mysqlinit.sql file using the
specified user.