VIA-1672 Immediately show pay button when registering for paid events.
Auto opens the collaped register pane when payment is required or
there is some other message.
Refactor the nested ifs for get_activity.
Fix dutch translation typo.
Reloads the page on success and reserve returns from the submit
ajax endpoint. This isn't ideal, but changing the javascript to include
showing the pay button and another message is too cumbersome.